Crate linera_alloy_signer_wallet

linera-alloy-signer-wallet

Local wallet implementations:

K256 private key

YubiHSM2

Features

keystore: enables Ethereum keystore functionality on the LocalWallet type.

mnemonic: enables BIP-39 mnemonic functionality for building LocalWallets.

yubihsm: enables Wallets with YubiHSM2 support.

Structs

Wallet

An Ethereum private-public key pair which can be used for signing messages.

Enums

WalletError

Error thrown by Wallet.

Type Aliases

LocalWallet

A wallet instantiated with a locally stored private key